entornos

+++ Entornos/Calculator/src/main/java/org/antonio/Calculator.java
@@ -10,8 +10,8 @@ public class Calculator {
         return a - b;
     }
 
-    public int multiply(int a, int b) {
-        return a * b;
+    public double multiply(double a, double b) {
+        return (a * b);
     }
 
     public float divide(int a, int b) {
